Slightly refactor Cecil tests
